arch: rename sparcv8leon3 to sparcv8leon
JIRA: RTOS-883
