+++ arch/sparcv8leon/Makefile
@@ -1,0 +1,9 @@
+#
+# Makefile for libphoenix/arch/sparcv8leon
+#
+# Copyright 2022 Phoenix Systems
+# Author: Lukasz Leczkowski
+#
+
+OBJS += $(addprefix $(PREFIX_O)arch/sparcv8leon/, syscalls.o jmp.o signal.o string.o reboot.o)
+CRT0_OBJS += $(addprefix $(PREFIX_O)arch/sparcv8leon/, crt0.o)
