+++ arch/sparcv8leon/reboot.c
@@ -18,11 +18,11 @@
 #include <sys/platform.h>
 
 #if defined(__CPU_GR716)
-#include <phoenix/arch/sparcv8leon3/gr716/gr716.h>
+#include <phoenix/arch/sparcv8leon/gr716/gr716.h>
 #elif defined(__CPU_GR712RC)
-#include <phoenix/arch/sparcv8leon3/gr712rc/gr712rc.h>
+#include <phoenix/arch/sparcv8leon/gr712rc/gr712rc.h>
 #elif defined(__CPU_GENERIC)
-#include <phoenix/arch/sparcv8leon3/generic/generic.h>
+#include <phoenix/arch/sparcv8leon/generic/generic.h>
 #else
 #error "Unsupported TARGET"
 #endif
