--- arch/sparcv8leon3/Makefile
@@ -1,9 +1,0 @@
-#
-# Makefile for libphoenix/arch/sparcv8leon3
-#
-# Copyright 2022 Phoenix Systems
-# Author: Lukasz Leczkowski
-#
-
-OBJS += $(addprefix $(PREFIX_O)arch/sparcv8leon3/, syscalls.o jmp.o signal.o string.o reboot.o)
-CRT0_OBJS += $(addprefix $(PREFIX_O)arch/sparcv8leon3/, crt0.o)
