+++ include/arch.h
@@ -13,7 +13,7 @@
 #elif defined(__riscv) && (__riscv_xlen == 64)
 #include <arch/riscv64/arch.h>
 #elif defined(__sparc__)
-#include <arch/sparcv8leon3/arch.h>
+#include <arch/sparcv8leon/arch.h>
 #else
 #error "unsupported architecture"
 #endif
