+++ include/arch/sparcv8leon/arch.h
@@ -3,7 +3,7 @@
  *
  * libphoenix
  *
- * Architecture dependent part (arch/sparcv8leon3)
+ * Architecture dependent part (arch/sparcv8leon)
  *
  * Copyright 2022 Phoenix Systems
  * Author: Lukasz Leczkowski
@@ -13,11 +13,11 @@
  * %LICENSE%
  */
 
-#ifndef _LIBPHOENIX_ARCH_SPARCV8LEON3_ARCH_H_
-#define _LIBPHOENIX_ARCH_SPARCV8LEON3_ARCH_H_
+#ifndef _LIBPHOENIX_ARCH_SPARCV8LEON_ARCH_H_
+#define _LIBPHOENIX_ARCH_SPARCV8LEON_ARCH_H_
 
-#define __ARCH_STDINT    <arch/sparcv8leon3/stdint.h>
-#define __ARCH_LIMITS    <arch/sparcv8leon3/limits.h>
+#define __ARCH_STDINT <arch/sparcv8leon/stdint.h>
+#define __ARCH_LIMITS <arch/sparcv8leon/limits.h>
 
 #define __MEMCPY
 #define __MEMCMP
@@ -58,7 +58,7 @@ static inline float __ieee754_sqrtf(float x)
 #else
 #define _PAGE_SIZE 0x1000
 #endif
-#define SIZE_PAGE  _Pragma("GCC warning \"'SIZE_PAGE' is deprecated. Use _PAGE_SIZE from arch.h or PAGE_SIZE from limits.h (POSIX only)\"") _PAGE_SIZE
+#define SIZE_PAGE _Pragma("GCC warning \"'SIZE_PAGE' is deprecated. Use _PAGE_SIZE from arch.h or PAGE_SIZE from limits.h (POSIX only)\"") _PAGE_SIZE
 
 #define __LIBPHOENIX_ARCH_TLS_SUPPORTED
